+++ src/variables/ConciergeVariable.php
@@ -1,0 +1,16 @@
+<?php
+namespace verbb\concierge\variables;
+
+use verbb\concierge\Concierge;
+
+class ConciergeVariable
+{
+    // Public Methods
+    // =========================================================================
+
+    public function getPlugin(): Concierge
+    {
+        return Concierge::$plugin;
+    }
+    
+}